finish functional happy flow

+++ app/src/main/java/com/example/minimoneybox/models/request/OneOffPaymentsRequest.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class OneOffPaymentsRequest(
     @SerializedName("Amount")
     @Expose
-    var amount: Int,
+    var amount: Float,
     @SerializedName("InvestorProductId")
     @Expose
     var investorProductId: Int
